+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - New subcommands
   - `print-dependabot`
   - `print-submodules-sh`
+  - `print-copy-files-from-senzing-install`
 
 ## [1.1.1] - 2021-02-22