+++ CHANGELOG.md
@@ -11,9 +11,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added in 1.2.0
 
 - New subcommands
+  - `print-copy-files-from-senzing-install`
   - `print-dependabot`
   - `print-submodules-sh`
-  - `print-copy-files-from-senzing-install`
 
 ## [1.1.1] - 2021-02-22