+++ CHANGELOG.md
@@ -26,6 +26,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added to 1.1.0
 
 - Initial subcommands:
-  - print-git-clone
-  - print-git-clone-mirror
-  - print-repository-names
+  - `print-git-clone`
+  - `print-git-clone-mirror`
+  - `print-repository-names`