+++ CHANGELOG.md
@@ -6,6 +6,15 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 [markdownlint](https://dlaa.me/markdownlint/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.2.0] - 2021-12-01
+
+### Added in 1.2.0
+
+- New subcommands
+  - `print-copy-files-from-senzing-install`
+  - `print-dependabot`
+  - `print-submodules-sh`
+
 ## [1.1.1] - 2021-02-22
 
 ### Changed in 1.1.1
@@ -17,6 +26,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added to 1.1.0
 
 - Initial subcommands:
-  - print-git-clone
-  - print-git-clone-mirror
-  - print-repository-names
+  - `print-git-clone`
+  - `print-git-clone-mirror`
+  - `print-repository-names`